+++ dart/geodata/lib/src/ogcapi_features/model.dart
@@ -1,3 +1,4 @@
+export 'model/ogc_feature_conformance.dart';
 export 'model/ogc_feature_item.dart';
 export 'model/ogc_feature_items.dart';
 export 'model/ogc_feature_service.dart';